move delegate paramter to its own module
