+++ qcodes/parameters/__init__.py
@@ -70,11 +70,12 @@ more specialized ones:
 
 from .array_parameter import ArrayParameter
 from .combined_parameter import CombinedParameter, combine
+from .delegate_parameter import DelegateParameter
 from .group_parameter import Group, GroupParameter
 from .grouped_parameter import DelegateGroup, DelegateGroupParameter, GroupedParameter
 from .multi_channel_instrument_parameter import MultiChannelInstrumentParameter
 from .multi_parameter import MultiParameter
-from .parameter import DelegateParameter, ManualParameter, Parameter
+from .parameter import ManualParameter, Parameter
 from .parameter_base import (
     ParamDataType,
     ParameterBase,
